+++ CHANGES.md
@@ -2,5 +2,8 @@
  - Oct. 2016
    - Added `Read-Choice` module
    - Added `Write-HtmlNode` module
+   - Added `SslWebBinding` module
+   - Added `AdministratorRole` module
+   - Added `Add-HostFileEntry` module
  - Aug. 2016
    - Added `Recycle` module